Introduce BrowserConnector
