+++ lib/puppeteer.rb
@@ -22,6 +22,7 @@ require 'puppeteer/aria_query_handler'
 require 'puppeteer/browser'
 require 'puppeteer/browser_context'
 require 'puppeteer/browser_runner'
+require 'puppeteer/browser_connector'
 require 'puppeteer/cdp_session'
 require 'puppeteer/chrome_target_manager'
 require 'puppeteer/connection'
